requirements.txt
ccxt
pandas
rich
python-dotenv
schedule
plotext
fastapi
uvicorn
jinja2
python-multipart
pandas-ta
xgboost
joblib
